# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Lightning POS Microservices API
  description: API documentation for Lightning POS microservices (Products and Orders).
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Products Service
  - url: http://localhost:3002
    description: Orders Service
  - url: http://localhost:3003
    description: Payment Service
paths:
  /v1/products/barcode/{barcode}:
    description: Products service endpoints
    get:
      summary: Get product by barcode
      tags:
        - Products
      parameters:
        - name: barcode
          in: path
          required: true
          schema:
            type: string
          description: The barcode of the product to retrieve
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
  /v1/orders:
    description: Orders service endpoints
    post:
      summary: Create a new order
      tags:
        - Orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderDto'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '500':
          description: Internal server error
  /v1/orders/{id}/status:
    patch:
      summary: Update order status
      tags:
        - Orders
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  $ref: '#/components/schemas/OrderStatus'
      responses:
        '200':
          description: Status updated
        '404':
          description: Order not found
  /v1/payments:
    post:
      summary: Process a payment
      tags:
        - Payments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcessPaymentDto'
      responses:
        '201':
          description: Payment processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
components:
  schemas:
    OrderStatus:
      type: string
      enum: [PENDING, PAID, CANCELLED]
    Product:
      type: object
      properties:
        id:
          type: string
        barcode:
          type: string
        name:
          type: string
        price:
          type: number
      required:
        - id
        - barcode
        - name
        - price
    OrderItem:
      type: object
      properties:
        productId:
          type: string
        quantity:
          type: integer
        price:
          type: number
      required:
        - productId
        - quantity
        - price
    Order:
      type: object
      properties:
        id:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        totalAmount:
          type: number
        createdAt:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/OrderStatus'
      required:
        - id
        - items
        - totalAmount
        - createdAt
        - status
    CreateOrderDto:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: string
              quantity:
                type: integer
              price:
                type: number
            required:
              - productId
              - quantity
              - price
      required:
        - items
    PaymentMethod:
      type: string
      enum: [CASH, CREDIT_CARD, QR_CODE]
    PaymentStatus:
      type: string
      enum: [SUCCESS, FAILED, PENDING]
    Payment:
      type: object
      properties:
        id:
          type: string
        orderId:
          type: string
        amount:
          type: number
        method:
          $ref: '#/components/schemas/PaymentMethod'
        status:
          $ref: '#/components/schemas/PaymentStatus'
        createdAt:
          type: string
          format: date-time
    ProcessPaymentDto:
      type: object
      properties:
        orderId:
          type: string
        amount:
          type: number
        method:
          $ref: '#/components/schemas/PaymentMethod'
      required:
        - orderId
        - amount
        - method